pyproject.toml
[project]
name = "infintree-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "authlib>=1.6.6",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.128.0",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
